requirements.txt
streamlit>=1.29.0
openai>=1.6.1
pandas>=2.2.0
rdflib>=7.0.0
networkx>=3.2.1
matplotlib>=3.8.2
plotly>=5.18.0
python-dotenv>=1.0.0
